// src/redux/features/critic/critic.types.ts
export type MyReviewItem = {
  criticDetail: {
    id: number;
    name: string;
    username: string;
    email: string;
    user_type: string;
    user_photo: string;
    user_token: string;
    website: string;
    country: {
      country_id: number;
      country_name: string;
      vat_price: number;
      country_badges: string;
    };
    about: string;
    facebook_url: string;
    twitter_url: string;
    gplus_url: string;
    instagram_url: string;
    linkedin_url: string;
    pinterest_url: string;
    verified: number;
    user_document_verified: number;
    country_badge: number;
    drop_status: string;
    city: string;
    zip_code: number;
    address: string;
    mobile: string;
    first_name: string;
    last_name: string;
    organization: string;
    designation: string;
    experience: string;
    image_path: string;
    total_review: string;
  };
  itemDetail: {
    item_id: number;
    price_status: string;
    price_percent: number;
    item_label_by_system: string;
    mainOwner: {
      id: number;
      name: string;
      username: string;
      email: string;
      user_type: string;
      user_photo: string;
      user_token: string;
      website: string;
      country: {
        country_id: number;
        country_name: string;
        vat_price: number;
        country_badges: string;
      };
      about: string;
      facebook_url: string;
      twitter_url: string;
      gplus_url: string;
      instagram_url: string;
      linkedin_url: string;
      pinterest_url: string;
      verified: number;
      user_document_verified: number;
      country_badge: number;
      drop_status: string;
      city: string;
      zip_code: number;
      address: string;
      mobile: string;
      first_name: string;
      last_name: string;
      organization: string;
      designation: string;
      experience: string;
      image_path: string;
    };
    item_token: string;
    item_name: string;
    item_slug: string;
    item_desc: string;
    item_shortdesc: string;
    item_thumbnail: string;
    item_preview: string;
    item_file: string;
    item_category: string;
    regular_price: number;
    item_tags: string;
    item_featured: string;
    size_id: string;
    medium_id: string;
    material_id: string;
    subject_id: string;
    is_adult: number;
    avg_critic_rating: number;
    avg_customer_rating: number;
    total_favorite: number;
    current_owner_id: number;
    image_path: string;
    total_review: string;
    total_review_product: string;
  };
  currentOwnerDetail: {
    id: number;
    name: string;
    username: string;
    email: string;
    user_type: string;
    user_photo: string;
    user_token: string;
    website: string;
    country: {
      country_id: number;
      country_name: string;
      vat_price: number;
      country_badges: string;
    };
    about: string;
    facebook_url: string;
    twitter_url: string;
    gplus_url: string;
    instagram_url: string;
    linkedin_url: string;
    pinterest_url: string;
    verified: number;
    user_document_verified: number;
    country_badge: number;
    drop_status: string;
    city: string;
    zip_code: number;
    address: string;
    mobile: string;
    first_name: string;
    last_name: string;
    organization: string;
    designation: string;
    experience: string;
    image_path: string;
  };
  rating: number;
  avg_critic_rating: number;
  avg_customer_rating: number;
  total_avg_rating: number;
  total_favorite: number;
  review: string;
  price: string;
  image_path: string;
  item_name: string;
  regular_price: string;
};

export type MyReviews = {
  current_page: number;
  data: MyReviewItem[];
  first_page_url: string;
  from: null;
  last_page: number;
  last_page_url: string;
  links: {
    url: null;
    label: string;
    active: boolean;
  }[];
  next_page_url: null;
  path: string;
  per_page: number;
  prev_page_url: null;
  to: null;
  total: number;
};